api/requirements.txt
anyio==3.4.0
asgiref==3.4.1
bcrypt==3.2.0
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
docopt==0.6.2
ecdsa==0.17.0
fastapi==0.70.0
h11==0.12.0
httptools==0.2.0
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
pipreqs==0.4.11
platformdirs==2.4.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.26.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
yarg==0.1.9
